// components/circular-gradiant-progressbar/hooks/interfaces.ts
type UseCircularGradiantProgressbarProps = {
  score: number;
  containerSize: number;
  strokeWidth: number;
  circleLength?: number;
  rotatesContainer?: number;
  maxValue?: number;
  strokeColor?: string[][];
  duration?: number;
  indicatorRadius?: number;
  indicatorStrokeWidth?: number;
  indicatorStrokeColor?: string;
};
export default UseCircularGradiantProgressbarProps;